requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
jinja2
jsonschema
httpx
python-dotenv
pyyaml
sqlparse
# Optional embeddings support
sentence-transformers
